ll-to-int

+++ SearchingAndSorting.cpp/MergeSort.cpp
@@ -5,17 +5,17 @@
 using namespace std;
 using namespace __gnu_pbds;
 #define w(x)                int x;cin>>x;while(x-->0)
-#define ll                  long long int
+#define int                  long long int
 #define mod                 1000000007
 #define arin(a)             for (auto &it : a) cin >> it;
-#define dbg(ar,st,ed)       for(ll i=st;i<=ed;i++)cout<<ar[i]<<" ";cout<<endl;
+#define dbg(ar,st,ed)       for(int i=st;i<=ed;i++)cout<<ar[i]<<" ";cout<<endl;
 #define endl               "\n"
-typedef tree<ll,null_type,less<ll>,rb_tree_tag,tree_order_statistics_node_update> ordered_set;
-typedef tree<ll,null_type,less_equal<ll>, rb_tree_tag,tree_order_statistics_node_update> ordered_multiset;
+typedef tree<int,null_type,less<int>,rb_tree_tag,tree_order_statistics_node_update> ordered_set;
+typedef tree<int,null_type,less_equal<int>, rb_tree_tag,tree_order_statistics_node_update> ordered_multiset;
 
-void merge(vector<ll>& ar,ll& l,ll& mid,ll& r){
-    vector<ll>temp;
-    ll i=l,j=mid+1;
+void merge(vector<int>& ar,int& l,int& mid,int& r){
+    vector<int>temp;
+    int i=l,j=mid+1;
     while(i<=mid && j<=r){
         if(ar[i]<=ar[j])temp.push_back(ar[i++]);
         else temp.push_back(ar[j++]);
@@ -23,14 +23,14 @@ void merge(vector<ll>& ar,ll& l,ll& mid,ll& r){
     while(i<=mid)temp.push_back(ar[i++]);
     while(j<=r)temp.push_back(ar[j++]);
 
-    for(ll i=l;i<=r;i++){
+    for(int i=l;i<=r;i++){
         ar[i]=temp[i-l];
     }
 }
 
-void mergeSort(vector<ll>& ar,ll l,ll r){
+void mergeSort(vector<int>& ar,int l,int r){
     if(l>=r)return;
-    ll mid = l+(r-l)/2;
+    int mid = l+(r-l)/2;
     mergeSort(ar,l,mid);
     mergeSort(ar,mid+1,r);
     merge(ar,l,mid,r);
@@ -42,11 +42,11 @@ void mergeSort(vector<ll>& ar,ll l,ll r){
 // Inversion Count Problem
 //Using Ordered_set
 void solve2(){
-    ll n;cin>>n;
-    vector<ll> ar(n);arin(ar);
-    ll ans=0;
+    int n;cin>>n;
+    vector<int> ar(n);arin(ar);
+    int ans=0;
     ordered_set set;
-    for(ll i=n-1;i>=0;i--){
+    for(int i=n-1;i>=0;i--){
         if(!set.empty()){
             ans+=set.order_of_key(ar[i]);
         }
@@ -56,20 +56,20 @@ void solve2(){
 }
 // Brute force O(n^2)
 void solve1(){
-    ll n;cin>>n;
-    vector<ll> ar(n);arin(ar);
-    ll ans=0;
-    for(ll i=0;i<n;i++){
-        for(ll j=i+1;j<n;j++){
+    int n;cin>>n;
+    vector<int> ar(n);arin(ar);
+    int ans=0;
+    for(int i=0;i<n;i++){
+        for(int j=i+1;j<n;j++){
             if(ar[i]>ar[j])ans+=1;
         }
     }
     cout<<ans<<endl;
 }
 //Using TwoPointer
-ll curAnsUsingTwoPointers(vector<ll>& ar,ll& l,ll& mid,ll& r){
-    ll ans=0;
-    ll i=l,j=mid+1;
+int curAnsUsingTwoPointers(vector<int>& ar,int& l,int& mid,int& r){
+    int ans=0;
+    int i=l,j=mid+1;
     while(i<=mid && j<=r){
         if(ar[i]<=ar[j]){
             ans+=((j-1)-mid);
@@ -84,13 +84,13 @@ ll curAnsUsingTwoPointers(vector<ll>& ar,ll& l,ll& mid,ll& r){
     return ans;
 }
 // Using Binary Search O(nlogn)
-ll curAnsUsingBs(vector<ll>& ar,ll& l,ll& mid,ll& r){
-    ll ans=0;
-    for(ll i=l;i<=mid;i++){
-        ll low=mid+1,high=r;
-        ll idx=-1;
+int curAnsUsingBs(vector<int>& ar,int& l,int& mid,int& r){
+    int ans=0;
+    for(int i=l;i<=mid;i++){
+        int low=mid+1,high=r;
+        int idx=-1;
         while(low<=high){
-            ll m = (low+high)/2;
+            int m = (low+high)/2;
             if(ar[m]<ar[i]){
                 idx=m;
                 low=m+1;
@@ -101,10 +101,10 @@ ll curAnsUsingBs(vector<ll>& ar,ll& l,ll& mid,ll& r){
     }
     return ans;
 }
-ll inversions(vector<ll>& ar,ll l,ll r){
+int inversions(vector<int>& ar,int l,int r){
     if(l>=r)return 0;
-    ll ans=0;
-    ll mid = (l+r)/2;
+    int ans=0;
+    int mid = (l+r)/2;
     ans+=inversions(ar,l,mid);
     ans+=inversions(ar,mid+1,r);
     ans+=curAnsUsingBs(ar,l,mid,r);
